[Network Commissioning] Remove old network commissioning driver
