+++ src/app/chip_data_model.gni
@@ -80,7 +80,6 @@ template("chip_data_model") {
       "${_app_root}/clusters/level-control/level-control.h",
       "${_app_root}/clusters/messaging-client/messaging-client.h",
       "${_app_root}/clusters/messaging-server/messaging-server.h",
-      "${_app_root}/clusters/network-commissioning-old/network-commissioning.h",
       "${_app_root}/clusters/on-off-server/on-off-server.h",
       "${_app_root}/clusters/scenes/scenes-tokens.h",
       "${_app_root}/clusters/scenes/scenes.h",
@@ -120,11 +119,6 @@ template("chip_data_model") {
     foreach(cluster, _cluster_sources) {
       if (cluster == "door-lock-server") {
         sources += [ "${_app_root}/clusters/${cluster}/door-lock-server.cpp" ]
-      } else if (cluster == "network-commissioning-old") {
-        sources += [
-          "${_app_root}/clusters/${cluster}/network-commissioning-ember.cpp",
-          "${_app_root}/clusters/${cluster}/network-commissioning-old.cpp",
-        ]
       } else if (cluster == "mode-select-server") {
         sources += [
           "${_app_root}/clusters/${cluster}/${cluster}.cpp",
